+++ publish/changeLog.md
@@ -1,4 +1,3 @@
 ### Feature
 
-- Add kg resource
-- Append image size parameter to wy image URL
+- Support extension store